\name{Rrrca-package}
\alias{Rrrca-package}
\alias{Rrrca}
\docType{package}
\title{
\packageTitle{Rrrca}
}
\description{
\packageDescription{Rrrca}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Rrrca}

Maintainer: \packageMaintainer{Rrrca}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Rrrca)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Rrrca}}
%\section{Documentation Index}{\packageIndices{Rrrca}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
